coverage: Extract safe FFI wrapper functions to `llvm_cov`
